Break layout and components into packages, individual files
