+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/layout/MeasurePolicy.kt
@@ -1,0 +1,7 @@
+package com.jakewharton.mosaic.layout
+
+import com.jakewharton.mosaic.layout.Measurable.MeasureScope
+
+public fun interface MeasurePolicy {
+	public fun MeasureScope.measure(measurables: List<Measurable>): MeasureResult
+}
