+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/layout/MeasureResult.kt
@@ -1,0 +1,7 @@
+package com.jakewharton.mosaic.layout
+
+public interface MeasureResult {
+	public val width: Int
+	public val height: Int
+	public fun placeChildren()
+}
