+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Row.kt
@@ -1,0 +1,28 @@
+@file:JvmName("Row")
+
+package com.jakewharton.mosaic.ui
+
+import androidx.compose.runtime.Composable
+import com.jakewharton.mosaic.layout.Layout
+import kotlin.jvm.JvmName
+
+@Composable
+public fun Row(content: @Composable () -> Unit) {
+	Layout(content, { "Row()" }) { measurables ->
+		var width = 0
+		var height = 0
+		val placeables = measurables.map { measurable ->
+			measurable.measure().also { placeable ->
+				width += placeable.width
+				height = maxOf(height, placeable.height)
+			}
+		}
+		layout(width, height) {
+			var x = 0
+			for (placeable in placeables) {
+				placeable.place(x, 0)
+				x += placeable.width
+			}
+		}
+	}
+}
