+++ samples/jest/src/main/kotlin/example/jest.kt
@@ -10,17 +10,17 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.Snapshot
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import com.jakewharton.mosaic.Color.Companion.Black
-import com.jakewharton.mosaic.Color.Companion.BrightBlack
-import com.jakewharton.mosaic.Color.Companion.Green
-import com.jakewharton.mosaic.Color.Companion.Red
-import com.jakewharton.mosaic.Color.Companion.Yellow
-import com.jakewharton.mosaic.Column
-import com.jakewharton.mosaic.Row
-import com.jakewharton.mosaic.Static
-import com.jakewharton.mosaic.Text
-import com.jakewharton.mosaic.TextStyle.Companion.Bold
 import com.jakewharton.mosaic.runMosaicBlocking
+import com.jakewharton.mosaic.ui.Color.Companion.Black
+import com.jakewharton.mosaic.ui.Color.Companion.BrightBlack
+import com.jakewharton.mosaic.ui.Color.Companion.Green
+import com.jakewharton.mosaic.ui.Color.Companion.Red
+import com.jakewharton.mosaic.ui.Color.Companion.Yellow
+import com.jakewharton.mosaic.ui.Column
+import com.jakewharton.mosaic.ui.Row
+import com.jakewharton.mosaic.ui.Static
+import com.jakewharton.mosaic.ui.Text
+import com.jakewharton.mosaic.ui.TextStyle.Companion.Bold
 import example.TestState.Fail
 import example.TestState.Pass
 import example.TestState.Running
